requirements.txt
streamlit
openai
langchain
chromadb
PyMuPDF
tiktoken
python-dotenv